fix(studio): move revalidateStudioPaths out of the use-server module
A "use server" file may only export async server actions, so the sync
revalidateStudioPaths helper broke the build. Moved it to a plain
server-only module (revalidate.ts) and updated imports.
